HDDS-3143. Rename silently ignored tests (#653)
